Solving variable power recipes  (#159)
* fix: Refactored Processor class into multiple smaller classes to improve maintenance

* fix: added average power

* fix: added new properties to Recipe interface for min and max power

* fix: added min and max power

* fix: code gardening

* fix: added converter variable power

* fix: code gardening

* fix: code gardening

* fix: code gardening

* fix: average power now working.

* fix: corrected tests

* fix: updated web data to v1.0.20

* fix: update CrystalShard name to Synthetic Power Shard merged from main

* Added automated miner

* fix: portable miner missing

* fix: parts test to handle miners

* fix: added quantum encoder!

* Added iron plate recipe test

* fix: iron plate test

* fix: added variable power test

* fix: cleanup of packages.json

* updated public web json file
