+++ web/src/config/config.ts
@@ -1,4 +1,4 @@
 export const config = {
   apiUrl: import.meta.env.VITE_ENV === 'dev' ? 'http://localhost:3001' : 'https://api.satisfactory-factories.app',
-  dataVersion: '1.0-19',
+  dataVersion: '1.0-20',
 }
